+++ app/src/main/java/io/github/christianjann/gitnotecje/ui/model/Grid.kt
@@ -25,7 +25,7 @@ enum class SortOrder {
 enum class NoteMinWidth(val size: Int) {
     C100(100),
     C150(150),
-    Default(200),
+    Default(150),
     C250(250),
     C300(300),
     C350(350),